Extract HelpTableFormatter and consistently use -s in source file names.
