+++ include/vcpkg/base/basic-checks.h
