+++ include/vcpkg/base/delayed-init.h
