+++ include/vcpkg/base/fwd/cmd-parser.h
@@ -1,0 +1,4 @@
+namespace vcpkg
+{
+    struct HelpTableFormatter;
+}
