+++ src/vcpkg-test/ci-baseline.cpp
@@ -1,6 +1,6 @@
 #include <catch2/catch.hpp>
 
-#include <vcpkg/base/basic_checks.h>
+#include <vcpkg/base/basic-checks.h>
 #include <vcpkg/base/parse.h>
 
 #include <vcpkg/build.h>
