+++ src/vcpkg-test/system.cpp
@@ -1,4 +1,4 @@
-#include <vcpkg/base/system_headers.h>
+#include <vcpkg/base/system-headers.h>
 
 #include <catch2/catch.hpp>
