+++ src/vcpkg/base/messages.cpp
@@ -1,6 +1,6 @@
 #include <vcpkg/base/json.h>
 #include <vcpkg/base/messages.h>
-#include <vcpkg/base/setup_messages.h>
+#include <vcpkg/base/setup-messages.h>
 #include <vcpkg/base/system.debug.h>
 #include <vcpkg/base/util.h>
