+++ src/vcpkg/base/system.process.cpp
@@ -1,4 +1,4 @@
-#include <vcpkg/base/system_headers.h>
+#include <vcpkg/base/system-headers.h>
 
 #include <vcpkg/base/checks.h>
 #include <vcpkg/base/chrono.h>
