+++ src/vcpkg/commands.regenerate.cpp
@@ -1,4 +1,4 @@
-#include <vcpkg/base/basic_checks.h>
+#include <vcpkg/base/basic-checks.h>
 #include <vcpkg/base/stringview.h>
 #include <vcpkg/base/util.h>
